Create named utilities out of vocabularies and avoid circular imports.
