+++ src/collective/easyform/browser/controlpanel.py
@@ -3,42 +3,17 @@ from collective.easyform import easyformMessageFactory as _
 from plone.app.registry.browser.controlpanel import ControlPanelFormWrapper
 from plone.app.registry.browser.controlpanel import RegistryEditForm
 from plone.registry.interfaces import IRegistry
-from plone.schemaeditor.interfaces import IFieldFactory
 from plone.z3cform import layout
 from z3c.form.browser.checkbox import CheckBoxFieldWidget
 from zope import schema
-from zope.component import getUtilitiesFor
 from zope.component import getUtility
-from zope.globalrequest import getRequest
-from zope.i18n import translate
-from zope.interface import implementer
 from zope.interface import Interface
-from zope.schema.interfaces import IVocabularyFactory
-from zope.schema.vocabulary import SimpleVocabulary
-
-import operator
 
 
 def getContent(self):
         return getUtility(IRegistry).forInterface(
             self.schema,
             prefix=self.schema_prefix)
-
-
-@implementer(IVocabularyFactory)
-def FieldsVocabularyFactory(context=None):
-    request = getRequest()
-    terms = []
-    for (field_id, factory) in getUtilitiesFor(IFieldFactory):
-        terms.append(
-            SimpleVocabulary.createTerm(
-                field_id,
-                factory.title,
-                translate(factory.title, context=request)
-            )
-        )
-    terms = sorted(terms, key=operator.attrgetter('title'))
-    return SimpleVocabulary(terms)
 
 
 class IEasyFormControlPanel(Interface):
@@ -60,7 +35,7 @@ class IEasyFormControlPanel(Interface):
                 default=u"Select the registry items you desire to modify"
             ),
             required=False,
-            vocabulary='collective.easyform.FieldsVocabulary',
+            vocabulary='easyform.SchemaEditorFields',
         ),
         default=[],
     )
