+++ src/collective/easyform/configure.zcml
@@ -43,4 +43,18 @@
         for="Products.CMFPlone.interfaces.IPloneSiteRoot"
         provides="Products.GenericSetup.interfaces.EXTENSION"
         />
+
+  <genericsetup:upgradeSteps
+      source="1000"
+      destination="1001"
+      profile="collective.easyform:default">
+    <genericsetup:upgradeDepends
+        title="Reload profile to register new vocabulary name"
+        description=""
+        import_profile="collective.easyform:default"
+        import_steps="plone.app.registry"
+        run_deps="false"
+        />
+  </genericsetup:upgradeSteps>
+
 </configure>
